Add async check for user account assignment
Introduced `IsPersonAssignedToUserAccountAsync` in `clsUser` and `clsUserData` to check if a person is linked to a user account.
Integrated this functionality into the `frmAddEditUser` form, updating the `btnNext_Click` event to handle the check asynchronously.
Added the `User_IsPersonAssignedToUserAccount` stored procedure to support this feature.

Enhanced `_ResetInformations` method and improved user feedback and error handling during user creation.
Included the new stored procedure in the
`ClinicWise.DataAccess.csproj` file.